+++ test/consecutive-glob-stars.js
@@ -1,9 +1,9 @@
 const t = require('tap')
-const {Minimatch, GLOBSTAR} = require('../')
+const { Minimatch, GLOBSTAR } = require('../')
 
 const patterns = []
 for (const a of ['**', '**/**', '**/**/**']) {
-  for (const b of['**', '**/**', '**/**/**']) {
+  for (const b of ['**', '**/**', '**/**/**']) {
     for (const c of ['**', '**/**', '**/**/**']) {
       patterns.push(`x/${a}/y/${b}/z/${c}`)
     }